@@ -25,22 +25,20 @@ use SatisPress\PackageType\ThemeBuilder;
  */
 final class PackageFactory {
 	/**
-	 * Release manager.
-	 *
-	 * @var ReleaseManager
-	 */
-	private $release_manager;
-
-	/**
 	 * Constructor.
 	 *
 	 * @since 0.3.0
 	 *
 	 * @param ReleaseManager $release_manager Release manager.
 	 */
-	public function __construct( ReleaseManager $release_manager ) {
-		$this->release_manager = $release_manager;
-	}
+	public function __construct(
+     /**
+      * Release manager.
+      */
+     private readonly ReleaseManager $release_manager
+ )
+ {
+ }
 
 	/**
 	 * Create a package builder.
@@ -50,14 +48,12 @@ final class PackageFactory {
 	 * @param string $package_type Package type.
 	 * @return PluginBuilder|ThemeBuilder|PackageBuilder Package builder instance.
 	 */
-	public function create( string $package_type ): PackageBuilder {
-		switch ( $package_type ) {
-			case 'plugin':
-				return new PluginBuilder( new Plugin(), $this->release_manager );
-			case 'theme':
-				return new ThemeBuilder( new Theme(), $this->release_manager );
-		}
-
-		return new PackageBuilder( new BasePackage(), $this->release_manager );
-	}
+	public function create(string $package_type): PackageBuilder
+ {
+     return match ($package_type) {
+         'plugin' => new PluginBuilder( new Plugin(), $this->release_manager ),
+         'theme' => new ThemeBuilder( new Theme(), $this->release_manager ),
+         default => new PackageBuilder( new BasePackage(), $this->release_manager ),
+     };
+ }
 }
